圆球

+++ XEngine/MeshBulid.h
@@ -12,8 +12,8 @@ public:
 	static MeshBulid* GetMeshBulid();
 	void CreateCyliner(float topr, float bottomr, UINT xrcount, UINT statckcount, float High, OUT vector<XVertx4> & vertxs,OUT vector<UINT16>& indess);
 	void CreateGrid(UINT width, UINT height,UINT m,UINT n,OUT vector<XVertx4> &invertxs,OUT vector<UINT16>& inindex);
-	void CreateSphere(UINT r,UINT m, UINT n, OUT vector<XVertx4> &invertxs, OUT vector<UINT16>& inindex);
-	void CreateBox(float w, float h, float d, OUT vector<XVertx4> &invertxs, OUT vector<UINT16>& inindex);
+	void CreateSphere(INT r,UINT m, UINT n, OUT vector<XVertx4> &invertxs, OUT vector<UINT>& inindex);
+	void CreateBox(float w, float h, float d, OUT vector<XVertx4> &invertxs, OUT vector<UINT>& inindex);
 private:
 	static MeshBulid* meshbulid;
